+++ internal/provider/building_block_definition_resource_test.go
@@ -697,6 +697,7 @@ func checkAzureDevopsPipelineImplementation() knownvalue.Check {
 	return knownvalue.MapExact(map[string]knownvalue.Check{
 		"project":     knownvalue.StringExact("MyProject"),
 		"pipeline_id": knownvalue.StringExact("42"),
+		"ref_name":    knownvalue.StringExact("refs/heads/main"),
 		"async":       knownvalue.Bool(false),
 		"integration_ref": knownvalue.MapExact(map[string]knownvalue.Check{
 			"uuid": KnownValueNotEmptyString(),